Refactored out test cases
